Decouple benchmarks from benchmark runner
