+++ bench/benchmarks/buffer.js
@@ -13,15 +13,16 @@ var Evented = require('../../js/util/evented');
 var config = require('../../js/util/config');
 var coordinates = require('../lib/coordinates');
 var formatNumber = require('../lib/format_number');
+var accessToken = require('../lib/access_token');
 
 var SAMPLE_COUNT = 10;
 
 module.exports = function run(options) {
-    config.ACCESS_TOKEN = options.accessToken;
+    config.ACCESS_TOKEN = accessToken;
 
     var evented = util.extend({}, Evented);
 
-    var stylesheetURL = 'https://api.mapbox.com/styles/v1/mapbox/streets-v9?access_token=' + options.accessToken;
+    var stylesheetURL = 'https://api.mapbox.com/styles/v1/mapbox/streets-v9?access_token=' + accessToken;
     ajax.getJSON(stylesheetURL, function(err, stylesheet) {
         if (err) return evented.fire('error', {error: err});
