+++ bench/benchmarks/geojson_setdata_small.js
@@ -5,6 +5,7 @@ var util = require('../../js/util/util');
 var formatNumber = require('../lib/format_number');
 var setDataPerf = require('../lib/set_data_perf');
 var setupGeoJSONMap = require('../lib/setup_geojson_map');
+var createMap = require('../lib/create_map');
 
 var featureCollection = {
     'type': 'FeatureCollection',
@@ -21,7 +22,7 @@ var featureCollection = {
 module.exports = function(options) {
     var evented = util.extend({}, Evented);
 
-    var map = options.createMap({
+    var map = createMap({
         width: 1024,
         height: 768,
         zoom: 5,
@@ -36,6 +37,7 @@ module.exports = function(options) {
 
         evented.fire('log', {message: 'loading small feature collection'});
         setDataPerf(source, 50, featureCollection, function(err, ms) {
+            map.remove();
             if (err) return evented.fire('error', {error: err});
             evented.fire('end', {message: formatNumber(ms) + ' ms', score: ms});
         });
