+++ bench/benchmarks/index.js
@@ -1,0 +1,10 @@
+window.mapboxglBenchmarks = {
+    'map-load': require('./map_load'),
+    'buffer': require('./buffer'),
+    'fps': require('./fps'),
+    'frame-duration': require('./frame_duration'),
+    'query-point': require('./query_point'),
+    'query-box': require('./query_box'),
+    'geojson-setdata-small': require('./geojson_setdata_small'),
+    'geojson-setdata-large': require('./geojson_setdata_large')
+};
