+++ bench/benchmarks/map_load.js
@@ -3,6 +3,7 @@
 var Evented = require('../../js/util/evented');
 var util = require('../../js/util/util');
 var formatNumber = require('../lib/format_number');
+var createMap = require('../lib/create_map');
 
 module.exports = function(options) {
     var evented = util.extend({}, Evented);
@@ -12,9 +13,10 @@ module.exports = function(options) {
     evented.fire('log', { message: 'Creating ' + mapsOnPage + ' maps' });
 
     var loaded = 0;
+    var maps = [];
     var start = Date.now();
     for (var i = 0; i < mapsOnPage; i++) {
-        var map = options.createMap({
+        var map = maps[i] = createMap({
             style: {
                 version: 8,
                 sources: {},
@@ -30,6 +32,9 @@ module.exports = function(options) {
     function onload () {
         if (++loaded >= mapsOnPage) {
             var duration = Date.now() - start;
+            for (var i = 0; i < maps.length; i++) {
+                maps[i].remove();
+            }
             evented.fire('end', {
                 message: formatNumber(duration) + ' ms, loaded ' + mapsOnPage + ' maps.',
                 score: duration
@@ -43,4 +48,3 @@ module.exports = function(options) {
 
     return evented;
 };
-
