+++ bench/benchmarks/query_point.js
@@ -2,6 +2,7 @@
 
 var Evented = require('../../js/util/evented');
 var util = require('../../js/util/util');
+var createMap = require('../lib/create_map');
 
 var width = 1024;
 var height = 768;
@@ -30,7 +31,7 @@ module.exports = function(options) {
 
     asyncSeries(zoomLevels.length, function(n, callback) {
         var zoomLevel = zoomLevels[zoomLevels.length - n];
-        var map = options.createMap({
+        var map = createMap({
             width: width,
             height: height,
             zoom: zoomLevel,
@@ -54,6 +55,7 @@ module.exports = function(options) {
                 zoomCount++;
                 callback();
             }, function() {
+                map.remove();
                 evented.fire('log', {
                     message: (zoomSum / zoomCount).toFixed(2) + ' ms at zoom ' + zoomLevel
                 });
