+++ bench/index.html
@@ -15,6 +15,7 @@
     <script src="https://unpkg.com/react@15.3.0/dist/react.js"></script>
     <script src="https://unpkg.com/react-dom@15.3.0/dist/react-dom.js"></script>
     <script src="/bench/index_generated.js"></script>
+    <script src="/bench/benchmarks_view_generated.js"></script>
 </body>
 
 </html>
