+++ bench/lib/access_token.js
@@ -1,0 +1,16 @@
+var accessToken = (
+    process.env.MapboxAccessToken ||
+    process.env.MAPBOX_ACCESS_TOKEN ||
+    getURLParameter('access_token') ||
+    localStorage.getItem('accessToken')
+);
+
+localStorage.setItem('accessToken', accessToken);
+
+module.exports = accessToken;
+
+function getURLParameter(name) {
+    var regexp = new RegExp('[?&]' + name + '=([^&#]*)', 'i');
+    var output = regexp.exec(window.location.href);
+    return output && output[1];
+}
