+++ package.json
@@ -89,11 +89,13 @@
     "watch-dev": "watchify js/mapbox-gl.js --debug --standalone mapboxgl --outfile dist/mapbox-gl-dev.js --verbose",
     "build-min": "browserify js/mapbox-gl.js --debug --transform unassertify --plugin [minifyify --map mapbox-gl.js.map --output dist/mapbox-gl.js.map] --standalone mapboxgl > dist/mapbox-gl.js && tap --no-coverage test/build/min.test.js",
     "build-token": "browserify debug/access_token.js --debug --transform envify > debug/access_token_generated.js",
-    "watch-bench": "node bench/download_data.js && watchify bench/index.js --plugin [minifyify --no-map] --transform [babelify --presets react] --transform unassertify --transform envify --outfile bench/index_generated.js --verbose",
+    "watch-benchmarks": "watchify bench/benchmarks/index.js --plugin [minifyify --no-map] --transform unassertify --transform envify --outfile bench/benchmarks/index_generated.js --verbose",
+    "watch-benchmarks-view": "watchify bench/benchmarks_view.js --plugin [minifyify --no-map] --transform [babelify --presets react] --transform unassertify --outfile bench/benchmarks_view_generated.js --verbose",
+    "download-benchmarks-data": "node bench/download_data.js",
     "start-server": "st --no-cache --localhost --port 9966 --index index.html .",
-    "start": "run-p build-token watch-dev watch-bench start-server",
+    "start": "run-p build-token watch-dev download-benchmarks-data watch-benchmarks watch-benchmarks-view start-server",
     "start-debug": "run-p build-token watch-dev start-server",
-    "start-bench": "run-p build-token watch-bench start-server",
+    "start-bench": "run-p build-token download-benchmarks-data watch-benchmarks watch-benchmarks-view start-server",
     "build-docs": "documentation build --github --format html --config documentation.yml --theme ./docs/_theme --output docs/api/",
     "build": "npm run build-docs # invoked by publisher when publishing docs on the mb-pages branch",
     "start-docs": "npm run build-min && npm run build-docs && jekyll serve --watch",
